[chore] fixed tests and docs
