+++ docs/mkdocs.yml
@@ -256,6 +256,8 @@ nav:
         - run list: 'cmd/flow/run/run-list.md'
         - run resubmit: 'cmd/flow/run/run-resubmit.md'
     - Power Platform (pp):
+      - dataverse:
+        - dataverse table list: 'cmd/pp/dataverse/dataverse-table-list.md'
       - environment:
         - environment list: 'cmd/pp/environment/environment-list.md'
       - gateway:
